refactor: move on-premises publishing payload construction
